requirements.txt
# Core
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Deployment
onnx>=1.14.0
onnxruntime>=1.15.0

# Config & Utilities
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0

# Data I/O
h5py>=3.9.0
mat73>=0.62

# Notebooks
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0

# Signal Processing
PyWavelets>=1.4.1
